% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtacSeqWorkshop-package.R
\docType{package}
\name{AtacSeqWorkshop-package}
\alias{AtacSeqWorkshop}
\alias{AtacSeqWorkshop-package}
\title{AtacSeqWorkshop: ATAC-Seq data analysis with Bioconductor}
\description{
This package contains subsets of real ATAC-Seq datasets and example
  analysis and quality control using Bioconductor tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/seandavi/AtacSeqWorkshop}
  \item Report bugs at \url{https://github.com/seandavi/AtacSeqWorkshop/issues/new}
}

}
\author{
\strong{Maintainer}: Sean Davis \email{seandavi@gmail.com}

Other contributors:
\itemize{
  \item Thomas Carroll \email{thomas.carroll@rockefeller.edu} [contributor]
}

}
\keyword{internal}
