% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtacSeqWorkshop-package.R
\docType{package}
\name{AtacSeqWorkshop-package}
\alias{AtacSeqWorkshop}
\alias{AtacSeqWorkshop-package}
\title{AtacSeqWorkshop: ATAC-Seq data analysis with Bioconductor}
\description{
This package contains subsets of real ATAC-Seq datasets and example
  analysis and quality control using Bioconductor tools.
}
\keyword{internal}
